ggml : introduce ggml-ext.h for staging new APIs
