+++ ggml/src/ggml-backend-meta.cpp
@@ -5,6 +5,9 @@
 #include "ggml-alloc.h"
 #include "ggml-cpp.h"
 
+// TODO: tmp
+#include "ggml-ext.h"
+
 #include <algorithm>
 #include <cassert>
 #include <cmath>
